deleted drawing.swift
