+++ RideFair/NavigationButton.swift
@@ -10,34 +10,6 @@ import UIKit
 
 class NavigationButton: UIButton {
 
-    override func draw(_ rect: CGRect) {
-      guard let context = UIGraphicsGetCurrentContext() else {
-        return
-      }
-      
-      // 1
-      let outerColor = UIColor(
-        hue: hue, saturation: saturation, brightness: brightness, alpha: 1.0)
-      let shadowColor = UIColor(red: 0.2, green: 0.2, blue: 0.2, alpha: 0.5)
-      
-      // 2
-      let outerMargin: CGFloat = 5.0
-      let outerRect = rect.insetBy(dx: outerMargin, dy: outerMargin)
-      // 3
-      let outerPath = createRoundedRectPath(for: outerRect, radius: 6.0)
-      
-      // 4
-      if state != .highlighted {
-        context.saveGState()
-        context.setFillColor(outerColor.cgColor)
-        context.setShadow(offset: CGSize(width: 0, height: 2),
-          blur: 3.0, color: shadowColor.cgColor)
-        context.addPath(outerPath)
-        context.fillPath()
-        context.restoreGState()
-      }
-    }
-    
     var hue: CGFloat {
       didSet {
         setNeedsDisplay()
@@ -66,6 +38,70 @@ class NavigationButton: UIButton {
       self.isOpaque = false
       self.backgroundColor = .clear
     }
+    
+    override func draw(_ rect: CGRect) {
+      guard let context = UIGraphicsGetCurrentContext() else {
+        return
+      }
+      
+      // 1
+      let outerColor = UIColor(
+        hue: hue, saturation: saturation, brightness: brightness, alpha: 1.0)
+      let shadowColor = UIColor(red: 0.2, green: 0.2, blue: 0.2, alpha: 0.5)
+      
+      // 2
+      let outerMargin: CGFloat = 5.0
+      let outerRect = rect.insetBy(dx: outerMargin, dy: outerMargin)
+      // 3
+      let outerPath = createRoundedRectPath(for: outerRect, radius: 6.0)
+      
+      // 4
+      if state != .highlighted {
+        context.saveGState()
+        context.setFillColor(outerColor.cgColor)
+        context.setShadow(offset: CGSize(width: 0, height: 2),
+          blur: 3.0, color: shadowColor.cgColor)
+        context.addPath(outerPath)
+        context.fillPath()
+        context.restoreGState()
+      }
+    }
+    
+    func createRoundedRectPath(for rect: CGRect, radius: CGFloat) -> CGMutablePath {
+        let path = CGMutablePath()
+        
+        // 1
+        let midTopPoint = CGPoint(x: rect.midX, y: rect.minY)
+        path.move(to: midTopPoint)
+        
+        // 2
+        let topRightPoint = CGPoint(x: rect.maxX, y: rect.minY)
+        let bottomRightPoint = CGPoint(x: rect.maxX, y: rect.maxY)
+        let bottomLeftPoint = CGPoint(x: rect.minX, y: rect.maxY)
+        let topLeftPoint = CGPoint(x: rect.minX, y: rect.minY)
+        
+        // 3
+        path.addArc(tangent1End: topRightPoint,
+          tangent2End: bottomRightPoint,
+          radius: radius)
+
+        path.addArc(tangent1End: bottomRightPoint,
+          tangent2End: bottomLeftPoint,
+          radius: radius)
+
+        path.addArc(tangent1End: bottomLeftPoint,
+          tangent2End: topLeftPoint,
+          radius: radius)
+
+        path.addArc(tangent1End: topLeftPoint,
+          tangent2End: topRightPoint,
+          radius: radius)
+
+        // 4
+        path.closeSubpath()
+        
+        return path
+    }
 
 //    override func draw(_ rect: CGRect) {
 //      guard let context = UIGraphicsGetCurrentContext() else {
@@ -81,22 +117,5 @@ class NavigationButton: UIButton {
 //      context.fill(bounds)
 //    }
     
-    func drawLinearGradient(
-      context: CGContext, rect: CGRect, startColor: CGColor, endColor: CGColor) {
-      // 1
-      let colorSpace = CGColorSpaceCreateDeviceRGB()
-      
-      // 2
-      let colorLocations: [CGFloat] = [0.0, 1.0]
-      
-      // 3
-      let colors: CFArray = [startColor, endColor] as CFArray
-      
-      // 4
-      let gradient = CGGradient(
-        colorsSpace: colorSpace, colors: colors, locations: colorLocations)!
-
-      // More to come...
-    }
     
 }
